rename some classes
