+++ KuleumBridge/app/src/main/java/com/KonDuckJoa/kuleumbridge/API/ApiGradeAll.java
@@ -1,15 +1,11 @@
 package com.KonDuckJoa.kuleumbridge.API;
 
-import android.content.Context;
 import android.os.AsyncTask;
-import android.widget.Toast;
 
 import com.KonDuckJoa.kuleumbridge.Common.CallBack;
 
 import org.json.JSONException;
 import org.json.JSONObject;
-
-import java.io.IOException;
 
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
@@ -17,35 +13,18 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 
-public class ApiLoginClass extends AsyncTask<String, String, Boolean> {
-    private String id, pwd, result;
-    private Response response;
-    private Context context;
+public class ApiGradeAll extends AsyncTask<String, String, Boolean> {
+    private String std_num, result;
     private CallBack cb;
 
-    public ApiLoginClass(String id, String pwd, Context context, CallBack cb)
+    public ApiGradeAll(String std_num, CallBack cb)
     {
-        this.context = context;
-        this.id = id;
-        this.pwd = pwd;
+        this.std_num = std_num;
         this.cb = cb;
     }
 
-
-    public String getResult() throws IOException {
-        return result;
-    }
-
-
     @Override
     protected void onPostExecute(Boolean success) {
-        if(success == null)
-        {
-            Toast.makeText(context, "네트워크가 불안정합니다. 다시 로그인해주세요.", Toast.LENGTH_SHORT).show();
-            cb.callback_fail();
-            return;
-        }
-
         super.onPostExecute(success);
 
         if(success)
@@ -54,24 +33,9 @@ public class ApiLoginClass extends AsyncTask<String, String, Boolean> {
         }
         else
         {
-            // json으로 받은 에러메세지에서 원하는 부분만 파싱하는 과정
-            JSONObject err_json = null;
-            try {
-                err_json = new JSONObject(result);
-
-                err_json = err_json.getJSONObject("ERRMSGINFO");
-
-                String msg = err_json.getString("ERRMSG");
-
-                // 토스트로 에러메세지 출력
-                Toast.makeText(context, msg, Toast.LENGTH_SHORT).show();
-
-                cb.callback_fail();
-            } catch (JSONException e) {
-                e.printStackTrace();
-            }
-
+            return;
         }
+
     }
 
     @Override
@@ -82,8 +46,7 @@ public class ApiLoginClass extends AsyncTask<String, String, Boolean> {
 
         JSONObject json = new JSONObject();
         try {
-            json.put("id", id);
-            json.put("pwd", pwd);
+            json.put("std_num", std_num);
         } catch (JSONException e) {
             e.printStackTrace();
         }
@@ -91,14 +54,15 @@ public class ApiLoginClass extends AsyncTask<String, String, Boolean> {
         // rest api 로그인 post로 보냄
         RequestBody body = RequestBody.create(JSON, json.toString());
         Request request = new Request.Builder()
-                .url("http://3.37.235.212:5000/login")
+                .url("http://3.37.235.212:5000/grade/all")
                 .addHeader("Connection", "close")
                 .post(body)
                 .build();
 
+        Response response;
+
         try {
             response = client.newCall(request).execute();
-
             result = response.body().string();
 
             if (result.contains("ERRMSGINFO")) {
@@ -106,11 +70,14 @@ public class ApiLoginClass extends AsyncTask<String, String, Boolean> {
             } else {
                 return true;
             }
-
         } catch (Exception e) {
             e.printStackTrace();
         }
 
-        return null;
+        return true;
+    }
+
+    public String getResult() {
+        return result;
     }
 }
