+++ KuleumBridge/app/src/main/java/com/KonDuckJoa/kuleumbridge/API/ApiGradeDetail.java
@@ -13,18 +13,21 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 
-public class ApiGradeNowClass extends AsyncTask<String, String, Boolean> {
-    private String std_num, result;
+public class ApiGradeDetail extends AsyncTask<String, String, Boolean> {
+    private String std_num, year, shtm, id, result;
     private CallBack cb;
 
-    public ApiGradeNowClass(String std_num, CallBack cb)
+    public ApiGradeDetail(String std_num, String year, String shtm, String id, CallBack cb)
     {
         this.std_num = std_num;
+        this.year = year;
+        this.shtm = shtm;
+        this.id = id;
         this.cb = cb;
     }
 
     @Override
-    protected void onPostExecute(Boolean success) {
+    protected void onPostExecute(java.lang.Boolean success) {
         super.onPostExecute(success);
 
         if(success)
@@ -39,7 +42,7 @@ public class ApiGradeNowClass extends AsyncTask<String, String, Boolean> {
     }
 
     @Override
-    protected Boolean doInBackground(String... strings) {
+    protected java.lang.Boolean doInBackground(String... strings) {
         final MediaType JSON = MediaType.parse("application/json; charset=utf-8");
 
         OkHttpClient client = new OkHttpClient();
@@ -47,6 +50,9 @@ public class ApiGradeNowClass extends AsyncTask<String, String, Boolean> {
         JSONObject json = new JSONObject();
         try {
             json.put("std_num", std_num);
+            json.put("year", year);
+            json.put("shtm", shtm);
+            json.put("sub_id", id);
         } catch (JSONException e) {
             e.printStackTrace();
         }
@@ -54,7 +60,7 @@ public class ApiGradeNowClass extends AsyncTask<String, String, Boolean> {
         // rest api 로그인 post로 보냄
         RequestBody body = RequestBody.create(JSON, json.toString());
         Request request = new Request.Builder()
-                .url("http://3.37.235.212:5000/grade/now")
+                .url("http://3.37.235.212:5000/grade/detail")
                 .addHeader("Connection", "close")
                 .post(body)
                 .build();
@@ -77,7 +83,8 @@ public class ApiGradeNowClass extends AsyncTask<String, String, Boolean> {
         return true;
     }
 
-    public String getResult() {
+    public java.lang.String getResult() {
         return result;
     }
 }
+
