+++ KuleumBridge/app/src/main/java/com/KonDuckJoa/kuleumbridge/API/ApiLogin.java
@@ -1,11 +1,15 @@
 package com.KonDuckJoa.kuleumbridge.API;
 
+import android.content.Context;
 import android.os.AsyncTask;
+import android.widget.Toast;
 
 import com.KonDuckJoa.kuleumbridge.Common.CallBack;
 
 import org.json.JSONException;
 import org.json.JSONObject;
+
+import java.io.IOException;
 
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
@@ -13,20 +17,35 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 
-public class ApiScheduleClass extends AsyncTask<String, String, Boolean> {
-    private String std_num, year, semester, result;
+public class ApiLogin extends AsyncTask<String, String, Boolean> {
+    private String id, pwd, result;
+    private Response response;
+    private Context context;
     private CallBack cb;
 
-    public ApiScheduleClass(String std_num, String year, String semester, CallBack cb)
+    public ApiLogin(String id, String pwd, Context context, CallBack cb)
     {
-        this.std_num = std_num;
-        this.year = year;
-        this.semester = semester;
+        this.context = context;
+        this.id = id;
+        this.pwd = pwd;
         this.cb = cb;
     }
 
+
+    public String getResult() throws IOException {
+        return result;
+    }
+
+
     @Override
     protected void onPostExecute(Boolean success) {
+        if(success == null)
+        {
+            Toast.makeText(context, "네트워크가 불안정합니다. 다시 로그인해주세요.", Toast.LENGTH_SHORT).show();
+            cb.callback_fail();
+            return;
+        }
+
         super.onPostExecute(success);
 
         if(success)
@@ -35,9 +54,24 @@ public class ApiScheduleClass extends AsyncTask<String, String, Boolean> {
         }
         else
         {
-            return;
-        }
+            // json으로 받은 에러메세지에서 원하는 부분만 파싱하는 과정
+            JSONObject err_json = null;
+            try {
+                err_json = new JSONObject(result);
 
+                err_json = err_json.getJSONObject("ERRMSGINFO");
+
+                String msg = err_json.getString("ERRMSG");
+
+                // 토스트로 에러메세지 출력
+                Toast.makeText(context, msg, Toast.LENGTH_SHORT).show();
+
+                cb.callback_fail();
+            } catch (JSONException e) {
+                e.printStackTrace();
+            }
+
+        }
     }
 
     @Override
@@ -48,9 +82,8 @@ public class ApiScheduleClass extends AsyncTask<String, String, Boolean> {
 
         JSONObject json = new JSONObject();
         try {
-            json.put("std_num", std_num);
-            json.put("year", year);
-            json.put("semester", semester);
+            json.put("id", id);
+            json.put("pwd", pwd);
         } catch (JSONException e) {
             e.printStackTrace();
         }
@@ -58,15 +91,14 @@ public class ApiScheduleClass extends AsyncTask<String, String, Boolean> {
         // rest api 로그인 post로 보냄
         RequestBody body = RequestBody.create(JSON, json.toString());
         Request request = new Request.Builder()
-                .url("http://3.37.235.212:5000/schedule")
+                .url("http://3.37.235.212:5000/login")
                 .addHeader("Connection", "close")
                 .post(body)
                 .build();
 
-        Response response;
-
         try {
             response = client.newCall(request).execute();
+
             result = response.body().string();
 
             if (result.contains("ERRMSGINFO")) {
@@ -74,14 +106,11 @@ public class ApiScheduleClass extends AsyncTask<String, String, Boolean> {
             } else {
                 return true;
             }
+
         } catch (Exception e) {
             e.printStackTrace();
         }
 
-        return true;
-    }
-
-    public String getResult() {
-        return result;
+        return null;
     }
 }
