+++ KuleumBridge/app/src/main/java/com/KonDuckJoa/kuleumbridge/Data/UserInfo.java
@@ -7,8 +7,8 @@ import org.json.JSONObject;
 
 import java.util.ArrayList;
 
-public class UserInfoClass{
-    private static UserInfoClass instance = new UserInfoClass();
+public class UserInfo {
+    private static UserInfo instance = new UserInfo();
 
     private String RESNO;       // 생년월일
     private String DEPT_TTNM;   // 소속단과대 및 학과
@@ -25,13 +25,13 @@ public class UserInfoClass{
     private StringBuilder grade_all_txt = new StringBuilder(); // 화면 상에서 보여지는 전체 성적 텍스트
     private StringBuilder grade_now_txt = new StringBuilder(); // 화면 상에서 보여지는 금학기 성적 텍스트
 
-    private UserInfoClass()
+    private UserInfo()
     {
         grade_all = new ArrayList<>();
         grade_now = new ArrayList<>();
     }
 
-    public static UserInfoClass getInstance()
+    public static UserInfo getInstance()
     {
         return instance;
     }
